refactor: sepparate level editing into its own class
